refactor(docker): move proxy transport state into proxy/state.rs
api.rs mixed three concerns: the proxy transport state machine
(ProxyState + EndpointReadiness), the router, and version-prefix
stripping. Move the transport state (and its tests) into the proxy
module where its collaborators live; api.rs keeps AppState + routing.
Also drop the empty system/image/network/volume route-group scaffolding
left over from the route de-enumeration — merging four empty routers
was noise.
