+++ app/arcbox-docker/src/api.rs
@@ -8,22 +8,15 @@
 
 use crate::handlers;
 use crate::proxy;
-use crate::proxy::{GuestConnector, GuestHttpClient};
+use crate::proxy::{GuestConnector, ProxyState};
 use crate::trace::trace_id_middleware;
 use arcbox_core::Runtime;
 use axum::extract::OriginalUri;
-use axum::http::{HeaderMap, Method, StatusCode};
 use axum::{
     Router, middleware,
     routing::{delete, post},
 };
-use bytes::Bytes;
-use http_body_util::BodyExt;
-use std::future::Future;
 use std::sync::Arc;
-use std::sync::Mutex;
-use std::sync::atomic::{AtomicU64, Ordering};
-use tokio::sync::Notify;
 
 /// Application state shared with handlers.
 #[derive(Clone)]
@@ -32,189 +25,6 @@ pub struct AppState {
     pub runtime: Arc<Runtime>,
     /// Guest proxy transport state.
     pub proxy: Arc<ProxyState>,
-}
-
-/// Guest proxy transport state shared by handlers.
-pub struct ProxyState {
-    connector: Arc<dyn GuestConnector>,
-    guest_http_client: GuestHttpClient,
-    endpoint_readiness: EndpointReadiness,
-}
-
-impl ProxyState {
-    fn new(connector: Arc<dyn GuestConnector>) -> Self {
-        Self {
-            guest_http_client: GuestHttpClient::new(Arc::clone(&connector)),
-            connector,
-            endpoint_readiness: EndpointReadiness::new(),
-        }
-    }
-
-    pub(crate) fn connector(&self) -> &dyn GuestConnector {
-        self.connector.as_ref()
-    }
-
-    pub(crate) fn client(&self) -> &GuestHttpClient {
-        &self.guest_http_client
-    }
-
-    /// Ensures guest dockerd is reachable at the Docker HTTP layer.
-    ///
-    /// The supplied `prepare_runtime` future owns the slow VM/agent/runtime
-    /// readiness path. This proxy state owns the cheaper HTTP `_ping`
-    /// verification and caches it until a transport failure invalidates it.
-    ///
-    /// `generation` is the System VM's current incarnation counter. When it
-    /// changes — the VM restarted (e.g. a backend switch) since the last call —
-    /// the cached readiness and pooled connections both point at the old VM, so
-    /// they are reset before verifying. Because this check is synchronous with
-    /// the request, it cannot race the restart the way an out-of-band event
-    /// watcher would.
-    pub(crate) async fn ensure_endpoint_verified<F>(
-        &self,
-        generation: u64,
-        prepare_runtime: F,
-    ) -> crate::error::Result<()>
-    where
-        F: Future<Output = crate::error::Result<()>>,
-    {
-        if self.endpoint_readiness.observe_generation(generation) {
-            // The readiness was already invalidated by `observe_generation`;
-            // also drop the pooled connections, which dialed the old VM.
-            self.guest_http_client.reset();
-        }
-        self.endpoint_readiness
-            .ensure_verified(prepare_runtime, || self.ping_guest())
-            .await
-    }
-
-    pub(crate) fn invalidate_endpoint(&self) {
-        self.endpoint_readiness.invalidate();
-    }
-
-    async fn ping_guest(&self) -> crate::error::Result<()> {
-        let response = proxy::proxy_to_guest_pooled(
-            &self.guest_http_client,
-            Method::GET,
-            "/_ping",
-            &HeaderMap::new(),
-            Bytes::new(),
-        )
-        .await?;
-
-        let status = response.status();
-        let body = BodyExt::collect(response.into_body())
-            .await
-            .map_err(|e| {
-                crate::error::DockerError::Server(format!(
-                    "failed to read guest docker _ping response: {e}"
-                ))
-            })?
-            .to_bytes();
-
-        if status == StatusCode::OK {
-            return Ok(());
-        }
-
-        Err(crate::error::DockerError::Server(format!(
-            "guest docker _ping returned {status}: {}",
-            String::from_utf8_lossy(&body).trim_end()
-        )))
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum EndpointReadinessState {
-    Unverified,
-    Verifying,
-    Verified,
-}
-
-struct EndpointReadiness {
-    state: Mutex<EndpointReadinessState>,
-    changed: Notify,
-    /// System VM incarnation this endpoint last verified against.
-    generation: AtomicU64,
-}
-
-impl EndpointReadiness {
-    fn new() -> Self {
-        Self {
-            state: Mutex::new(EndpointReadinessState::Unverified),
-            changed: Notify::new(),
-            generation: AtomicU64::new(0),
-        }
-    }
-
-    /// Records the current VM incarnation and, when it differs from the last
-    /// one seen (the System VM restarted in between), invalidates the cached
-    /// readiness and reports the change so the caller can drop stale pooled
-    /// connections too.
-    fn observe_generation(&self, generation: u64) -> bool {
-        if self.generation.swap(generation, Ordering::AcqRel) == generation {
-            return false;
-        }
-        self.invalidate();
-        true
-    }
-
-    async fn ensure_verified<Prepare, Verify, VerifyFuture>(
-        &self,
-        prepare_runtime: Prepare,
-        verify_endpoint: Verify,
-    ) -> crate::error::Result<()>
-    where
-        Prepare: Future<Output = crate::error::Result<()>>,
-        Verify: FnOnce() -> VerifyFuture,
-        VerifyFuture: Future<Output = crate::error::Result<()>>,
-    {
-        loop {
-            let wait_for_change = {
-                let mut state = self.state.lock().unwrap_or_else(|e| e.into_inner());
-                match *state {
-                    EndpointReadinessState::Verified => return Ok(()),
-                    EndpointReadinessState::Unverified => {
-                        *state = EndpointReadinessState::Verifying;
-                        None
-                    }
-                    EndpointReadinessState::Verifying => Some(self.changed.notified()),
-                }
-            };
-
-            if let Some(wait_for_change) = wait_for_change {
-                wait_for_change.await;
-                continue;
-            }
-
-            let result = async {
-                prepare_runtime.await?;
-                verify_endpoint().await
-            }
-            .await;
-
-            let mut state = self.state.lock().unwrap_or_else(|e| e.into_inner());
-            *state = if result.is_ok() {
-                EndpointReadinessState::Verified
-            } else {
-                EndpointReadinessState::Unverified
-            };
-            self.changed.notify_waiters();
-            return result;
-        }
-    }
-
-    fn invalidate(&self) {
-        let mut state = self.state.lock().unwrap_or_else(|e| e.into_inner());
-        if *state != EndpointReadinessState::Unverified {
-            *state = EndpointReadinessState::Unverified;
-            self.changed.notify_waiters();
-        }
-    }
-
-    #[cfg(test)]
-    fn state(&self) -> EndpointReadinessState {
-        *self.state.lock().unwrap_or_else(|e| e.into_inner())
-    }
 }
 
 /// Creates the Docker API router with all endpoints.
@@ -292,16 +102,8 @@ fn strip_version_prefix(path: &str) -> Option<&str> {
 
 fn api_routes() -> Router<AppState> {
     Router::new()
-        .merge(system_routes())
         .merge(container_routes())
-        .merge(build_routes())
-        .merge(image_routes())
-        .merge(network_routes())
-        .merge(volume_routes())
-}
-
-fn system_routes() -> Router<AppState> {
-    Router::new()
+        .route("/build", post(handlers::build_image))
 }
 
 fn container_routes() -> Router<AppState> {
@@ -328,28 +130,9 @@ fn container_routes() -> Router<AppState> {
         )
 }
 
-fn build_routes() -> Router<AppState> {
-    Router::new().route("/build", post(handlers::build_image))
-}
-
-fn image_routes() -> Router<AppState> {
-    Router::new()
-}
-
-fn network_routes() -> Router<AppState> {
-    Router::new()
-}
-
-fn volume_routes() -> Router<AppState> {
-    Router::new()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::error::DockerError;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use tokio::time::{Duration, sleep};
 
     #[test]
     fn strip_standard_version_prefix() {
@@ -419,206 +202,5 @@ mod tests {
         let req = strip_api_version_prefix(req);
         assert_eq!(req.uri().path(), "/containers/abc/start");
         assert!(req.extensions().get::<OriginalUri>().is_none());
-    }
-
-    #[tokio::test]
-    async fn readiness_transitions_unverified_to_verified_after_success() {
-        let readiness = EndpointReadiness::new();
-        let prepared = Arc::new(AtomicUsize::new(0));
-        let verified = Arc::new(AtomicUsize::new(0));
-
-        assert_eq!(readiness.state(), EndpointReadinessState::Unverified);
-
-        let prepared_current = Arc::clone(&prepared);
-        let verified_current = Arc::clone(&verified);
-        readiness
-            .ensure_verified(
-                async move {
-                    prepared_current.fetch_add(1, Ordering::Relaxed);
-                    Ok::<(), DockerError>(())
-                },
-                || async move {
-                    verified_current.fetch_add(1, Ordering::Relaxed);
-                    Ok::<(), DockerError>(())
-                },
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(readiness.state(), EndpointReadinessState::Verified);
-        assert_eq!(prepared.load(Ordering::Relaxed), 1);
-        assert_eq!(verified.load(Ordering::Relaxed), 1);
-    }
-
-    #[tokio::test]
-    async fn readiness_keeps_verified_state_on_cache_hit() {
-        let readiness = EndpointReadiness::new();
-        let prepared = Arc::new(AtomicUsize::new(0));
-        let verified = Arc::new(AtomicUsize::new(0));
-
-        let prepared_first = Arc::clone(&prepared);
-        let verified_first = Arc::clone(&verified);
-        readiness
-            .ensure_verified(
-                async move {
-                    prepared_first.fetch_add(1, Ordering::Relaxed);
-                    Ok::<(), DockerError>(())
-                },
-                || async move {
-                    verified_first.fetch_add(1, Ordering::Relaxed);
-                    Ok::<(), DockerError>(())
-                },
-            )
-            .await
-            .unwrap();
-
-        let prepared_second = Arc::clone(&prepared);
-        let verified_second = Arc::clone(&verified);
-        readiness
-            .ensure_verified(
-                async move {
-                    prepared_second.fetch_add(1, Ordering::Relaxed);
-                    Ok::<(), DockerError>(())
-                },
-                || async move {
-                    verified_second.fetch_add(1, Ordering::Relaxed);
-                    Ok::<(), DockerError>(())
-                },
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(prepared.load(Ordering::Relaxed), 1);
-        assert_eq!(verified.load(Ordering::Relaxed), 1);
-    }
-
-    #[tokio::test]
-    async fn readiness_transitions_back_to_unverified_after_failure() {
-        let readiness = EndpointReadiness::new();
-        let prepared = Arc::new(AtomicUsize::new(0));
-        let verified = Arc::new(AtomicUsize::new(0));
-
-        let prepared_current = Arc::clone(&prepared);
-        let verified_current = Arc::clone(&verified);
-        let err = readiness
-            .ensure_verified(
-                async move {
-                    prepared_current.fetch_add(1, Ordering::Relaxed);
-                    Ok::<(), DockerError>(())
-                },
-                || async move {
-                    verified_current.fetch_add(1, Ordering::Relaxed);
-                    Err::<(), DockerError>(DockerError::Server("ping failed".into()))
-                },
-            )
-            .await
-            .unwrap_err();
-
-        assert!(err.to_string().contains("ping failed"));
-        assert_eq!(readiness.state(), EndpointReadinessState::Unverified);
-        assert_eq!(prepared.load(Ordering::Relaxed), 1);
-        assert_eq!(verified.load(Ordering::Relaxed), 1);
-    }
-
-    #[tokio::test]
-    async fn readiness_invalidation_transitions_verified_to_unverified() {
-        let readiness = EndpointReadiness::new();
-        let verified = Arc::new(AtomicUsize::new(0));
-
-        for _ in 0..2 {
-            let verified_current = Arc::clone(&verified);
-            readiness
-                .ensure_verified(async { Ok::<(), DockerError>(()) }, || async move {
-                    verified_current.fetch_add(1, Ordering::Relaxed);
-                    Ok::<(), DockerError>(())
-                })
-                .await
-                .unwrap();
-            readiness.invalidate();
-        }
-
-        assert_eq!(verified.load(Ordering::Relaxed), 2);
-        assert_eq!(readiness.state(), EndpointReadinessState::Unverified);
-    }
-
-    #[tokio::test]
-    async fn readiness_generation_change_invalidates_verified_state() {
-        let readiness = EndpointReadiness::new();
-
-        // Verify against the initial incarnation (generation 0).
-        readiness
-            .ensure_verified(async { Ok::<(), DockerError>(()) }, || async {
-                Ok::<(), DockerError>(())
-            })
-            .await
-            .unwrap();
-        assert_eq!(readiness.state(), EndpointReadinessState::Verified);
-
-        // Re-observing the same incarnation is a cache hit — no invalidation.
-        assert!(!readiness.observe_generation(0));
-        assert_eq!(readiness.state(), EndpointReadinessState::Verified);
-
-        // A new incarnation (the VM restarted) drops the cached verification.
-        assert!(readiness.observe_generation(7));
-        assert_eq!(readiness.state(), EndpointReadinessState::Unverified);
-
-        // Stable once recorded.
-        assert!(!readiness.observe_generation(7));
-    }
-
-    #[tokio::test]
-    async fn readiness_serializes_concurrent_verification() {
-        let readiness = Arc::new(EndpointReadiness::new());
-        let prepared = Arc::new(AtomicUsize::new(0));
-        let verified = Arc::new(AtomicUsize::new(0));
-
-        let first_readiness = Arc::clone(&readiness);
-        let first_prepared = Arc::clone(&prepared);
-        let first_verified = Arc::clone(&verified);
-        let first = tokio::spawn(async move {
-            first_readiness
-                .ensure_verified(
-                    async move {
-                        first_prepared.fetch_add(1, Ordering::Relaxed);
-                        sleep(Duration::from_millis(20)).await;
-                        Ok::<(), DockerError>(())
-                    },
-                    || async move {
-                        first_verified.fetch_add(1, Ordering::Relaxed);
-                        sleep(Duration::from_millis(20)).await;
-                        Ok::<(), DockerError>(())
-                    },
-                )
-                .await
-        });
-
-        while readiness.state() != EndpointReadinessState::Verifying {
-            sleep(Duration::from_millis(1)).await;
-        }
-
-        let second_readiness = Arc::clone(&readiness);
-        let second_prepared = Arc::clone(&prepared);
-        let second_verified = Arc::clone(&verified);
-        let second = tokio::spawn(async move {
-            second_readiness
-                .ensure_verified(
-                    async move {
-                        second_prepared.fetch_add(1, Ordering::Relaxed);
-                        Ok::<(), DockerError>(())
-                    },
-                    || async move {
-                        second_verified.fetch_add(1, Ordering::Relaxed);
-                        Ok::<(), DockerError>(())
-                    },
-                )
-                .await
-        });
-
-        first.await.unwrap().unwrap();
-        second.await.unwrap().unwrap();
-
-        assert_eq!(readiness.state(), EndpointReadinessState::Verified);
-        assert_eq!(prepared.load(Ordering::Relaxed), 1);
-        assert_eq!(verified.load(Ordering::Relaxed), 1);
     }
 }
