+++ app/arcbox-docker/src/proxy/mod.rs
@@ -8,6 +8,7 @@ mod fallback;
 mod forward;
 mod headers;
 mod session;
+mod state;
 mod upgrade;
 mod upload;
 mod uri;
@@ -18,6 +19,7 @@ pub use forward::{
     proxy_to_guest, proxy_to_guest_pooled, proxy_to_guest_stream, proxy_to_guest_stream_pooled,
 };
 pub use session::GuestHttpClient;
+pub use state::ProxyState;
 pub use upgrade::proxy_with_upgrade;
 pub use upload::proxy_streaming_upload;
